Server Functionality
Implemented splitting of string
Implemented for loop to check for message identifier
