+++ nbproject/private/private.xml
@@ -3,6 +3,8 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
+            <file>file:/C:/Users/N0639343/Desktop/JavaProject-master/src/javaproject/readingFile.java</file>
+            <file>file:/C:/Users/N0639343/Desktop/JavaProject-master/src/javaproject/NTU.png</file>
             <file>file:/C:/Users/N0639343/Desktop/JavaProject-master/src/javaproject/NTUSpotify.java</file>
             <file>file:/C:/Users/N0639343/Desktop/JavaProject-master/src/javaproject/SocketClient.java</file>
             <file>file:/C:/Users/N0639343/Desktop/JavaProject-master/src/javaproject/MultipleSocketServer.java</file>
