+++ src/javaproject/NTUSpotify.form
@@ -82,7 +82,7 @@
                           </Group>
                           <Group type="102" alignment="0" attributes="0">
                               <EmptySpace max="-2" attributes="0"/>
-                              <Component id="label" min="-2" pref="257" max="-2" attributes="0"/>
+                              <Component id="ConnectionLabel" min="-2" pref="257" max="-2" attributes="0"/>
                           </Group>
                       </Group>
                       <EmptySpace pref="233" max="32767" attributes="0"/>
@@ -93,7 +93,7 @@
               <Group type="103" groupAlignment="0" attributes="0">
                   <Group type="102" alignment="0" attributes="0">
                       <EmptySpace max="-2" attributes="0"/>
-                      <Component id="label" min="-2" pref="14" max="-2" attributes="0"/>
+                      <Component id="ConnectionLabel" min="-2" pref="14" max="-2" attributes="0"/>
                       <EmptySpace type="separate" max="-2" attributes="0"/>
                       <Group type="103" groupAlignment="3" attributes="0">
                           <Component id="UsernameLabel1" alignment="3" min="-2" max="-2" attributes="0"/>
@@ -160,7 +160,7 @@
                 <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="PasswordFieldActionPerformed"/>
               </Events>
             </Component>
-            <Component class="javax.swing.JLabel" name="label">
+            <Component class="javax.swing.JLabel" name="ConnectionLabel">
             </Component>
           </SubComponents>
         </Container>
