+++ src/javaproject/SocketClient.java
@@ -23,13 +23,17 @@ public class SocketClient {
          InetAddress address = InetAddress.getByName(host);
          Socket connection = new Socket(address, port);
          
+         //Write the output to a buffered output and then to the writer
          BufferedOutputStream bos = new BufferedOutputStream(connection.getOutputStream());
          OutputStreamWriter osw = new OutputStreamWriter(bos, "US-ASCII");
          
-         String message = username + " " + password + (char) 13;
+         //Message to send over server
+         String message = username + "-" + password + "-" + "readFile" + (char) 13;
+         //Writing the message and flushing the server
          osw.write(message);
          osw.flush();
-                  
+         
+         //Close the connection
          connection.close();
          isCorrect = "correct";
      }
